requirements.txt
Django==5.0.2
djangorestframework==3.14.0
python-dotenv==1.0.1
google-generativeai==0.3.2
Pillow==11.1.0
django-cors-headers==4.3.1
python-jose==3.3.0
bcrypt==4.1.2
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
pymongo==4.6.1
dj-database-url==2.1.0